% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd-data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin Market Price (USD)}
\format{
An xts object with 1354 observations and 1 variable on
2019-01-02 / 2023-10-28 containing:
\describe{
  \item{Data}{double [1354, 1]}
  \item{Index}{POSIXct,POSIXt [1354] (TZ: "UTC")}
}
}
\source{
<https://www.blockchain.com/explorer/charts/market-price>
}
\usage{
bitcoin
}
\description{
The average USD market price across major bitcoin exchanges.
}
\examples{
if (!requireNamespace("ggplot2", quietly = TRUE)) utils::install.packages(
  "ggplot2", repos = "https://cloud.r-project.org", quiet = TRUE
)

ggplot2::ggplot(bitcoin, ggplot2::aes(x = date, y = price)) +
  ggplot2::geom_line()
}
\keyword{datasets}
