% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_glm.R
\name{cost_glm}
\alias{cost_glm}
\title{Logistic regression.}
\usage{
cost_glm(data, family = "binomial")
}
\arguments{
\item{data}{TODO}

\item{family}{TODO}
}
\value{
TODO
}
\description{
Logistic regression.
}
