% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_lasso.R
\name{cost_lasso}
\alias{cost_lasso}
\title{Logistic regression.}
\usage{
cost_lasso(data, lambda, family = "gaussian")
}
\arguments{
\item{data}{TODO}

\item{lambda}{TODO}

\item{family}{TODO}
}
\value{
TODO
}
\description{
Logistic regression.
}
