% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_lasso_update.R
\name{cost_lasso_update}
\alias{cost_lasso_update}
\title{Solve logistic regression using Gradient Descent Extension to the
multivariate case}
\usage{
cost_lasso_update(data, coef, cum_coef, cmatrix, lambda)
}
\arguments{
\item{data}{TODO}

\item{coef}{TODO}

\item{cum_coef}{TODO}

\item{cmatrix}{TODO}

\item{lambda}{TODO}
}
\value{
TODO
}
\description{
Solve logistic regression using Gradient Descent Extension to the
multivariate case
}
