% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_logistic_update.R
\name{cost_logistic_update}
\alias{cost_logistic_update}
\title{Solve logistic regression using Gradient Descent Extension to the
multivariate case}
\usage{
cost_logistic_update(data_new, coef, cum_coef, cmatrix, epsilon = 1e-10)
}
\arguments{
\item{data_new}{TODO}

\item{coef}{TODO}

\item{cum_coef}{TODO}

\item{cmatrix}{TODO}

\item{epsilon}{TODO}
}
\value{
TODO
}
\description{
Solve logistic regression using Gradient Descent Extension to the
multivariate case
}
