% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_poisson_update.R
\name{cost_poisson_update}
\alias{cost_poisson_update}
\title{Solve poisson regression using Gradient Descent Extension to the
multivariate case}
\usage{
cost_poisson_update(
  data,
  coef,
  cum_coef,
  cmatrix,
  epsilon = 0.001,
  G = 10^10,
  L = -20,
  H = 20
)
}
\arguments{
\item{data}{TODO}

\item{coef}{TODO}

\item{cum_coef}{TODO}

\item{cmatrix}{TODO}

\item{epsilon}{TODO}

\item{G}{TODO}

\item{L}{TODO}

\item{H}{TODO}
}
\value{
TODO
}
\description{
Solve poisson regression using Gradient Descent Extension to the
multivariate case
}
