% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_update.R
\name{cost_update}
\alias{cost_update}
\title{Solve logistic regression using Gradient Descent Extension to the
multivariate case}
\usage{
cost_update(
  data,
  theta_hat,
  theta_sum,
  hessian,
  family,
  momentum,
  momentum_coef,
  args_list
)
}
\arguments{
\item{data}{TODO}

\item{theta_hat}{TODO}

\item{theta_sum}{TODO}

\item{hessian}{TODO}

\item{family}{TODO}

\item{momentum}{TODO}

\item{momentum_coef}{TODO}

\item{args_list}{TODO}
}
\value{
TODO
}
\description{
Solve logistic regression using Gradient Descent Extension to the
multivariate case
}
