% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_update_gradient}
\alias{cost_update_gradient}
\title{Function to calculate the gradient at the current data.}
\usage{
cost_update_gradient(data, theta, family)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta}{Estimated theta from the previous iteration.}

\item{family}{Family of the model.}
}
\value{
Gradient at the current data.
}
\description{
Function to calculate the gradient at the current data.
}
