% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_update_hessian.R
\name{cost_update_hessian}
\alias{cost_update_hessian}
\title{Function to calculate the Hessian matrix at the current data.}
\usage{
cost_update_hessian(data, theta, hessian, family, G = NULL)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta}{Estimated theta from the previous iteration.}

\item{hessian}{Hessian matrix from the previous iteration.}

\item{family}{Family of the model.}
}
\value{
Hessian at the current data.
}
\description{
Function to calculate the Hessian matrix at the current data.
}
