% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{CP}
\alias{CP}
\title{Dynamic Programming with Pruning and Gradient Descent}
\usage{
CP(data, beta, B = 10, trim = 0.025)
}
\arguments{
\item{data}{TODO}

\item{beta}{TODO}

\item{B}{TODO}

\item{trim}{TODO}
}
\value{
TODO
}
\description{
Dynamic Programming with Pruning and Gradient Descent
}
