% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Generate data from logistic regression models with change-points}
\usage{
data_gen(n, d, true.coef, true.cp.loc, Sigma)
}
\arguments{
\item{n}{TODO}

\item{d}{TODO}

\item{true.coef}{TODO}

\item{true.cp.loc}{TODO}

\item{Sigma}{TODO}
}
\value{
TODO
}
\description{
Generate data from logistic regression models with change-points
}
