% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_wrappers.R
\name{fastcpd.arima}
\alias{fastcpd.arima}
\alias{fastcpd_arima}
\title{Find change points efficiently in ARIMA(p, d, q) models}
\usage{
fastcpd.arima(data, order = 0, ...)

fastcpd_arima(data, order = 0, ...)
}
\arguments{
\item{data}{A numeric vector, a matrix, a data frame or a time series object.}

\item{order}{A vector of length three specifying the order of the ARIMA
model.}

\item{...}{Other arguments passed to \code{\link{fastcpd}}, for example,
\code{segment_count}. One special argument can be passed here is
\code{include.mean}, which is a logical value indicating whether the
mean should be included in the model. The default value is \code{TRUE}.}
}
\value{
A class \code{fastcpd} object.
}
\description{
\code{fastcpd_arima} and \code{fastcpd.arima} are
wrapper functions of \code{\link{fastcpd}} to find change points in
ARIMA(p, d, q) models. The function is similar to \code{\link{fastcpd}}
except that the data is by default a one-column matrix or univariate vector
and thus a formula is not required here.
}
\examples{
\donttest{
set.seed(1)
n <- 1000
x <- rep(0, n + 3)
for (i in 1:600) {
  x[i + 3] <- 0.6 * x[i + 2] - 0.2 * x[i + 1] + 0.1 * x[i] + rnorm(1, 0, 3)
}
for (i in 601:1000) {
  x[i + 3] <- 0.3 * x[i + 2] + 0.4 * x[i + 1] + 0.2 * x[i] + rnorm(1, 0, 3)
}
result <- fastcpd.arima(
  x[3 + seq_len(n)],
  c(3, 0, 0),
  include.mean = FALSE,
  trim = 0,
  beta = (3 + 1 + 1) * log(n) / 2 * 5
)
summary(result)
plot(result)
}
}
