% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{fastcpd_exponential}
\alias{fastcpd_exponential}
\alias{fastcpd.exponential}
\title{Find change points efficiently in exponentially distributed data}
\usage{
fastcpd_exponential(data, ...)

fastcpd.exponential(data, ...)
}
\arguments{
\item{data}{A matrix, a data frame or a vector.}

\item{...}{Other arguments passed to \code{\link[=fastcpd]{fastcpd()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=fastcpd_exponential]{fastcpd_exponential()}} and \code{\link[=fastcpd.exponential]{fastcpd.exponential()}} are
wrapper functions of \code{\link[=fastcpd]{fastcpd()}} to find changes in the rate of
exponentially distributed data, i.e. mean change under exponentially
distributed noise (cf. \code{changepoint::cpt.meanvar} with
\code{test.stat = "Exponential"}). The function is similar to
\code{\link[=fastcpd]{fastcpd()}} except that the data is by default a matrix or data frame or
a vector with each row / element as an observation and thus a formula is
not required here.
}
\examples{
set.seed(1)
data <- matrix(c(
  rexp(300, rate = 1),
  rexp(400, rate = 0.1),
  rexp(300, rate = 1)
))
system.time(result <- fastcpd.exponential(data))
summary(result)
plot(result)
}
\seealso{
\code{\link[=fastcpd]{fastcpd()}}
}
