% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_wrappers.R
\name{fastcpd.ma}
\alias{fastcpd.ma}
\alias{fastcpd_ma}
\title{Find change points efficiently in MA(q) models}
\usage{
fastcpd.ma(data, order = 0, ...)

fastcpd_ma(data, order = 0, ...)
}
\arguments{
\item{data}{A numeric vector, a matrix, a data frame or a time series object.}

\item{order}{A positive integer or a vector of length three with the first
two elements being zeros specifying the order of the MA model.}

\item{...}{Other arguments passed to \code{\link{fastcpd}}, for example,
\code{segment_count}. One special argument can be passed here is
\code{include.mean}, which is a logical value indicating whether the
mean should be included in the model. The default value is \code{TRUE}.}
}
\value{
A class \code{fastcpd} object.
}
\description{
\code{fastcpd_ma} and \code{fastcpd.ma} are
wrapper functions of \code{\link{fastcpd}} to find change points in
MA(q) models. The function is similar to \code{\link{fastcpd}}
except that the data is by default a one-column matrix or univariate vector
and thus a formula is not required here.
}
\examples{
\donttest{
set.seed(1)
n <- 400
p <- 4
time_series <- rep(0, n)
time_series[1:200] <- arima.sim(
  n = 200, model = list(ma = c(-5 / 3, 11 / 12, -5 / 12, 1 / 6), sd = 0.1)
)
time_series[201:400] <- arima.sim(
  n = 200, model = list(ma = c(-4 / 3, 7 / 9, -16 / 27, 4 / 27), sd = 0.3)
)
result <- fastcpd.ma(time_series, 4, include.mean = FALSE, trim = 0)
}
}
