% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastcpd_vanilla}
\alias{fastcpd_vanilla}
\title{Vanilla PELT implementation.
This function is not meant to be called directly by the user.}
\usage{
fastcpd_vanilla(
  data,
  beta,
  segment_count,
  trim,
  momentum_coef,
  k,
  family,
  epsilon,
  min_prob,
  winsorise_minval,
  winsorise_maxval,
  p,
  cost,
  cp_only
)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{beta}{Initial cost value.}

\item{segment_count}{Number of segments for initial guess.}

\item{trim}{Trimming for the boundary change points.}

\item{momentum_coef}{Momentum coefficient to be applied to each update.}

\item{k}{Function on number of epochs in SGD.}

\item{family}{Family of the model. Can be "binomial", "poisson", "lasso" or
"gaussian". If not provided, the user must specify the cost function and
its gradient (and Hessian).}

\item{epsilon}{Epsilon to avoid numerical issues. Only used for binomial and
poisson.}

\item{min_prob}{Minimum probability to avoid numerical issues. Only used for
poisson.}

\item{winsorise_minval}{Minimum value to be winsorised. Only used for
poisson.}

\item{winsorise_maxval}{Maximum value to be winsorised. Only used for
poisson.}

\item{p}{Number of parameters to be estimated.}

\item{cost}{Cost function to be used. If not specified, the default is
the negative log-likelihood for the corresponding family.}

\item{cp_only}{Whether to return only the change points or with the cost
values for each segment.}
}
\value{
Change points and corresponding cost values.
}
\description{
Vanilla PELT implementation.
This function is not meant to be called directly by the user.
}
