% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_log_lik.R
\name{neg_log_lik}
\alias{neg_log_lik}
\title{Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case}
\usage{
neg_log_lik(data, b, family, lambda = NULL)
}
\arguments{
\item{data}{TODO}

\item{b}{TODO}

\item{family}{TODO}

\item{lambda}{TODO}
}
\value{
TODO
}
\description{
Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case
}
