% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_log_likelihood.R
\name{negative_log_likelihood}
\alias{negative_log_likelihood}
\title{Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case}
\usage{
negative_log_likelihood(data, b, family, lambda = NULL)
}
\arguments{
\item{data}{TODO}

\item{b}{TODO}

\item{family}{TODO}

\item{lambda}{TODO}
}
\value{
TODO
}
\description{
Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case
}
