% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{plot,fastcpd-method}
\alias{plot,fastcpd-method}
\title{Plot the data and the change points for a \code{fastcpd} object}
\usage{
\S4method{plot}{fastcpd}(x)
}
\arguments{
\item{x}{\code{fastcpd} object.}
}
\description{
Plot the data and the change points for a \code{fastcpd} object
}
