% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{print,fastcpd-method}
\alias{print,fastcpd-method}
\title{Print the call and the change points for a \code{fastcpd} object}
\usage{
\S4method{print}{fastcpd}(x)
}
\arguments{
\item{x}{\code{fastcpd} object.}
}
\description{
Print the call and the change points for a \code{fastcpd} object
}
