% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd-class.R
\name{print.fastcpd}
\alias{print.fastcpd}
\alias{print,fastcpd-method}
\title{Print the call and the change points for a \code{fastcpd} object}
\usage{
\method{print}{fastcpd}(x, ...)

\S4method{print}{fastcpd}(x, ...)
}
\arguments{
\item{x}{\code{fastcpd} object.}

\item{...}{Ignored.}
}
\description{
Print the call and the change points for a \code{fastcpd} object
}
