% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{show,fastcpd-method}
\alias{show,fastcpd-method}
\title{Show the available methods for a \code{fastcpd} object}
\usage{
\S4method{show}{fastcpd}(object)
}
\arguments{
\item{object}{\code{fastcpd} object.}
}
\description{
Show the available methods for a \code{fastcpd} object
}
