% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{summary,fastcpd-method}
\alias{summary,fastcpd-method}
\title{Show the summary of a \code{fastcpd} object}
\usage{
\S4method{summary}{fastcpd}(object)
}
\arguments{
\item{object}{\code{fastcpd} object.}
}
\description{
Show the summary of a \code{fastcpd} object
}
