% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd-class.R
\name{summary.fastcpd}
\alias{summary.fastcpd}
\alias{summary,fastcpd-method}
\title{Show the summary of a \code{fastcpd} object}
\usage{
\method{summary}{fastcpd}(object, ...)

\S4method{summary}{fastcpd}(object, ...)
}
\arguments{
\item{object}{\code{fastcpd} object.}

\item{...}{Ignored.}
}
\description{
Show the summary of a \code{fastcpd} object
}
