% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd-class.R
\name{summary.fastcpd}
\alias{summary.fastcpd}
\alias{summary,fastcpd-method}
\title{Show the summary of a \code{fastcpd} object}
\usage{
\method{summary}{fastcpd}(object, ...)

\S4method{summary}{fastcpd}(object, ...)
}
\arguments{
\item{object}{\code{fastcpd} object.}

\item{...}{Ignored.}
}
\value{
Return a (temporarily) invisible copy of the \code{fastcpd} object.
    Called primarily for printing the summary of the model including the
    call, the change points, the cost values and the estimated parameters.
}
\description{
Show the summary of a \code{fastcpd} object
}
