% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd-data.R
\docType{data}
\name{transcriptome}
\alias{transcriptome}
\title{Transcription Profiling of 57 Human Bladder Carcinoma Samples}
\format{
A data frame with 2215 rows and 43 variables:
\describe{
  \item{columns}{Corresponding to the 43 individuals}
}
}
\source{
<https://www.ebi.ac.uk/biostudies/arrayexpress/studies/E-TABM-147>

<https://github.com/cran/ecp/tree/master/data>
}
\usage{
transcriptome
}
\description{
Transcriptome analysis of 57 bladder carcinomas on Affymetrix HG-U95A and
HG-U95Av2 microarrays
}
\examples{
result <- fastcpd.mean(transcriptome[, 8], trim = 0.005)
summary(result)
plot(result)
}
\keyword{datasets}
