% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_estimation.R
\name{variance_lm}
\alias{variance_lm}
\alias{variance.lm}
\title{Variance estimation for linear models with change points}
\usage{
variance_lm(data, block_size = NULL, p = NULL)

variance.lm(data, block_size = NULL, p = NULL)
}
\arguments{
\item{data}{A matrix or a data frame with the response variable as the first
column.}

\item{block_size}{The size of the blocks to use for variance estimation.}

\item{p}{The number of predictors in the model.}
}
\value{
A numeric value representing the variance.
}
\description{
Estimate the variance for each block and then take the average.
}
\examples{
if (!requireNamespace("mvtnorm", quietly = TRUE)) utils::install.packages(
  "mvtnorm", repos = "https://cloud.r-project.org", quiet = TRUE
)

set.seed(1)
n <- 300
p <- 4
x <- mvtnorm::rmvnorm(n, rep(0, p), diag(p))
theta_0 <- rbind(c(1, 3.2, -1, 0), c(-1, -0.5, 2.5, -2), c(0.8, 0, 1, 2))
y <- c(
  x[1:100, ] \%*\% theta_0[1, ] + rnorm(100, 0, 3),
  x[101:200, ] \%*\% theta_0[2, ] + rnorm(100, 0, 3),
  x[201:300, ] \%*\% theta_0[3, ] + rnorm(100, 0, 3)
)
(sigma2 <- variance.lm(cbind(y, x)))
}
