% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_estimation.R
\name{variance_median}
\alias{variance_median}
\alias{variance.median}
\title{Variance estimation for median change models}
\usage{
variance_median(data)

variance.median(data)
}
\arguments{
\item{data}{A vector of data points.}
}
\value{
A numeric value representing the variance.
}
\description{
Implement Rice estimator.
}
\examples{
(sigma2 <- variance.median(well_log))
}
