% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_update.R
\name{cost_update}
\alias{cost_update}
\title{Update the cost values for the segmentation.}
\usage{
cost_update(
  data,
  theta_hat,
  theta_sum,
  hessian,
  tau,
  i,
  k,
  family,
  momentum,
  momentum_coef,
  args_list
)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta_hat}{Estimated theta from the previous iteration.}

\item{theta_sum}{Sum of estimated theta from the previous iteration.}

\item{hessian}{Hessian matrix from the previous iteration.}

\item{tau}{Start of the current segment.}

\item{i}{Index of the current data in the whole data set.}

\item{k}{Number of epochs in SGD.}

\item{family}{Family of the model.}

\item{momentum}{Momentum from the previous iteration.}

\item{momentum_coef}{Momentum coefficient to be applied to the current
momentum.}

\item{args_list}{Arguments to be passed to the model.}
}
\value{
A list containing new values of \code{theta_hat}, \code{theta_sum},
  \code{hessian}, and \code{momentum}.s
}
\description{
Update the cost values for the segmentation.
}
