% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_update}
\alias{cost_update}
\title{Update the cost values for the segmentation.
This function is not meant to be called directly by the user.}
\usage{
cost_update(
  data,
  theta_hat,
  theta_sum,
  hessian,
  tau,
  i,
  k,
  family,
  momentum,
  momentum_coef,
  epsilon,
  min_prob,
  winsorise_minval,
  winsorise_maxval,
  lambda,
  cost_gradient,
  cost_hessian
)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta_hat}{Estimated theta from the previous iteration.}

\item{theta_sum}{Sum of estimated theta from the previous iteration.}

\item{hessian}{Hessian matrix from the previous iteration.}

\item{tau}{Start of the current segment.}

\item{i}{Index of the current data in the whole data set.}

\item{k}{Number of epochs in SGD.}

\item{family}{Family of the model.}

\item{momentum}{Momentum from the previous iteration.}

\item{momentum_coef}{Momentum coefficient to be applied to the current
momentum.}

\item{epsilon}{Epsilon to avoid numerical issues. Only used for binomial and
poisson.}

\item{min_prob}{Minimum probability to avoid numerical issues. Only used for
poisson.}

\item{winsorise_minval}{Minimum value to be winsorised. Only used for
poisson.}

\item{winsorise_maxval}{Maximum value to be winsorised. Only used for
poisson.}

\item{lambda}{Lambda for L1 regularization. Only used for lasso.}

\item{cost_gradient}{Gradient for custom cost function.}

\item{cost_hessian}{Hessian for custom cost function.}
}
\value{
A list containing new values of \code{theta_hat}, \code{theta_sum},
  \code{hessian}, and \code{momentum}.
}
\description{
Update the cost values for the segmentation.
This function is not meant to be called directly by the user.
}
\keyword{internal}
