% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_update_hessian}
\alias{cost_update_hessian}
\title{Function to calculate the Hessian matrix at the current data.
This function is not meant to be called directly by the user.}
\usage{
cost_update_hessian(data, theta, family, min_prob)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta}{Estimated theta from the previous iteration.}

\item{family}{Family of the model.}

\item{min_prob}{Minimum probability to avoid numerical issues.}
}
\value{
Hessian at the current data.
}
\description{
Function to calculate the Hessian matrix at the current data.
This function is not meant to be called directly by the user.
}
