% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{CP}
\alias{CP}
\title{Dynamic Programming with Pruning and Gradient Descent}
\usage{
CP(
  data,
  beta,
  segment_count = 10,
  trim = 0.025,
  momentum_coef = 0,
  family,
  ...
)
}
\arguments{
\item{data}{TODO}

\item{beta}{TODO}

\item{segment_count}{TODO}

\item{trim}{TODO}

\item{family}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Dynamic Programming with Pruning and Gradient Descent
}
