% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_vanilla.R
\name{CP_vanilla}
\alias{CP_vanilla}
\title{Dynamic Programming with Pruning}
\usage{
CP_vanilla(data, beta, cost = cost_glm, family)
}
\arguments{
\item{data}{TODO}

\item{beta}{TODO}

\item{cost}{TODO}

\item{family}{TODO}
}
\value{
TODO
}
\description{
Dynamic Programming with Pruning
}
