% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{detect_var}
\alias{detect_var}
\alias{fastcpd_var}
\alias{fastcpd.var}
\title{Find change points efficiently in VAR(\eqn{p}) models}
\usage{
detect_var(data, order = 0, ...)

fastcpd_var(data, order = 0, ...)

fastcpd.var(data, order = 0, ...)
}
\arguments{
\item{data}{A matrix, a data frame or a time series object.}

\item{order}{A positive integer specifying the order of the VAR model.}

\item{...}{Other arguments passed to \code{\link[=detect]{detect()}}, for example,
\code{segment_count}.}
}
\value{
A \link[=fastcpd-class]{fastcpd} object.
}
\description{
\code{\link[=detect_var]{detect_var()}} and \code{\link[=fastcpd.var]{fastcpd.var()}} are
wrapper functions of \code{\link[=detect]{detect()}} to find change points in
VAR(\eqn{p}) models. The function is similar to \code{\link[=detect]{detect()}}
except that the data is by default a matrix with row as an observation
and thus a formula is not required here. Pass the raw, unlagged series;
its response dimension is inferred from the number of columns. The old
Python-only pre-lagged code{p_response}/code{p.response} form is not part
of the portable interface.
}
\examples{
set.seed(1)
n <- 300
p <- 2
theta_1 <- matrix(c(-0.3, 0.6, -0.5, 0.4, 0.2, 0.2, 0.2, -0.2), nrow = p)
theta_2 <- matrix(c(0.3, -0.4, 0.1, -0.5, -0.5, -0.2, -0.5, 0.2), nrow = p)
x <- matrix(0, n + 2, p)
for (i in 1:200) {
  x[i + 2, ] <- theta_1 \%*\% c(x[i + 1, ], x[i, ]) + rnorm(p, 0, 1)
}
for (i in 201:n) {
  x[i + 2, ] <- theta_2 \%*\% c(x[i + 1, ], x[i, ]) + rnorm(p, 0, 1)
}
result <- detect_var(x, 2)
summary(result)
}
\seealso{
\code{\link[=detect]{detect()}}
}
