% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\name{fastcpd}
\alias{fastcpd}
\title{Dynamic Programming with Pruning and Gradient Descent}
\usage{
fastcpd(
  data,
  beta,
  segment_count = 10,
  trim = 0.025,
  momentum_coef = 0,
  sgd_k = 3,
  family,
  cost = negative_log_likelihood,
  ...
)
}
\arguments{
\item{data}{TODO}

\item{beta}{TODO}

\item{segment_count}{TODO}

\item{trim}{TODO}

\item{momentum_coef}{TODO}

\item{sgd_k}{TODO}

\item{family}{TODO}

\item{cost}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Dynamic Programming with Pruning and Gradient Descent
}
