% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\docType{class}
\name{fastcpd}
\alias{fastcpd}
\title{An S4 class to store the three data.frames created with \link{lnt_read}}
\usage{
fastcpd(
  data,
  beta,
  segment_count = 10,
  trim = 0.025,
  momentum_coef = 0,
  sgd_k = 3,
  family = NULL,
  epsilon = 1e-10,
  min_prob = 10^10,
  winsorise_minval = -20,
  winsorise_maxval = 20,
  p = NULL,
  cost = negative_log_likelihood,
  cost_gradient = cost_update_gradient,
  cost_hessian = cost_update_hessian
)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{beta}{Initial cost value.}

\item{segment_count}{Number of segments for initial guess.}

\item{trim}{Trimming for the boundary change points.}

\item{momentum_coef}{Momentum coefficient to be applied to each update.}

\item{sgd_k}{Number of epochs in SGD.}

\item{family}{Family of the model. Can be "binomial", "poisson", "lasso" or
"gaussian". If not provided, the user must specify the cost function and
its gradient (and Hessian).}

\item{epsilon}{Epsilon to avoid numerical issues. Only used for binomial and
poisson.}

\item{min_prob}{Minimum probability to avoid numerical issues. Only used for
poisson.}

\item{winsorise_minval}{Minimum value to be winsorised. Only used for
poisson.}

\item{winsorise_maxval}{Maximum value to be winsorised. Only used for
poisson.}

\item{cost}{Cost function to be used. If not specified, the default is
the negative log-likelihood for the corresponding family.}

\item{cost_gradient}{Gradient for custom cost function.}

\item{cost_hessian}{Hessian for custom cost function.}

\item{lambda}{Lambda for L1 regularization. Only used for lasso.}
}
\value{
Change points and corresponding cost values.
}
\description{
This S4 class stores the output from \link{lnt_read}. Just like a spreadsheet
with multiple worksheets, an fastcpd object consist of three data.frames
which you can select using \code{@}. This object class is intended to be an
intermediate container. As it stores articles and paragraphs in two separate
data.frames, nested in an S4 object, the relevant text data is stored twice
in almost the same format. This has the advantage, that there is no need to
use special characters, such as "\\n" to indicate a new paragraph. However,
it makes the files rather big when you save them directly. They should thus
usually be subsetted using \code{@} or converted to a different format using
\link{lnt_convert}.
}
\section{Slots}{

\describe{
\item{\code{meta}}{The metadata of the articles read in.}

\item{\code{articles}}{The article texts and respective IDs.}

\item{\code{paragraphs}}{The paragraphs (if the data.frame exists) and respective
article and paragraph IDs.}
}}

