% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd.R
\docType{package}
\name{fastcpd}
\alias{fastcpd}
\title{fastcpd: A package for finding change points in an efficient way}
\usage{
fastcpd(
  formula = y ~ . - 1,
  data,
  beta = NULL,
  segment_count = 10,
  trim = 0.025,
  momentum_coef = 0,
  k = function(x) 0,
  family = NULL,
  epsilon = 1e-10,
  min_prob = 10^10,
  winsorise_minval = -20,
  winsorise_maxval = 20,
  p = NULL,
  cost = negative_log_likelihood,
  cost_gradient = cost_update_gradient,
  cost_hessian = cost_update_hessian,
  cp_only = TRUE
)

fastcpd(
  formula = y ~ . - 1,
  data,
  beta = NULL,
  segment_count = 10,
  trim = 0.025,
  momentum_coef = 0,
  k = function(x) 0,
  family = NULL,
  epsilon = 1e-10,
  min_prob = 10^10,
  winsorise_minval = -20,
  winsorise_maxval = 20,
  p = NULL,
  cost = negative_log_likelihood,
  cost_gradient = cost_update_gradient,
  cost_hessian = cost_update_hessian,
  cp_only = TRUE
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{data}{A data frame containing the data to be segmented.}

\item{beta}{Initial cost value.}

\item{segment_count}{Number of segments for initial guess.}

\item{trim}{Trimming for the boundary change points.}

\item{momentum_coef}{Momentum coefficient to be applied to each update.}

\item{k}{Function on number of epochs in SGD.}

\item{family}{Family of the model. Can be "binomial", "poisson", "lasso" or
"gaussian". If not provided, the user must specify the cost function and
its gradient (and Hessian).}

\item{epsilon}{Epsilon to avoid numerical issues. Only used for binomial and
poisson.}

\item{min_prob}{Minimum probability to avoid numerical issues. Only used for
poisson.}

\item{winsorise_minval}{Minimum value to be winsorised. Only used for
poisson.}

\item{winsorise_maxval}{Maximum value to be winsorised. Only used for
poisson.}

\item{p}{Number of parameters to be estimated.}

\item{cost}{Cost function to be used. If not specified, the default is
the negative log-likelihood for the corresponding family.}

\item{cost_gradient}{Gradient for custom cost function.}

\item{cost_hessian}{Hessian for custom cost function.}

\item{cp_only}{Whether to return only the change points or with the cost
values for each segment.}
}
\value{
A class \code{fastcpd} object.
}
\description{
The fastcpd package provides a function \code{\link{fastcpd}} to find change
points in a data set. The function is based on the paper "Sequential Gradient
Descent and Quasi-Newton’s Method for Change-Point Analysis" by
Xianyang Zhang and Trisha Dawn.

This S4 class stores the output from \link{fastcpd}. A fastcpd object consist
of several slots including the call to \link{fastcpd}, the data used, the
family of the model, the change points, the cost values, the residuals, the
estimated parameters and a boolean indicating whether the model was fitted
with only change points or with change points and parameters, which you can
select using \code{@}.
}
\section{Slots}{

\describe{
\item{\code{call}}{The call to \link{fastcpd}.}

\item{\code{data}}{The data used.}

\item{\code{family}}{The family of the model.}

\item{\code{cp_set}}{The change points.}

\item{\code{cost_values}}{The cost values for each segment.}

\item{\code{residuals}}{The residuals for each segment.}

\item{\code{thetas}}{The estimated parameters for each segment.}

\item{\code{cp_only}}{A boolean indicating whether the model was fitted with only
change points or with change points and parameters.}
}}

\section{fastcpd functions}{

The fastcpd functions ...
}

\examples{
library(fastcpd)
# Binomial
x <- matrix(rnorm(1500, 0, 1), ncol = 5)
theta <- rbind(rnorm(5, 0, 1), rnorm(5, 2, 1))
y <- c(
  rbinom(125, 1, 1 / (1 + exp(-x[1:125, ] \%*\% theta[1, ]))),
  rbinom(175, 1, 1 / (1 + exp(-x[126:300, ] \%*\% theta[2, ])))
)
result <- fastcpd(y ~ . - 1, data.frame(y = y, x = x), family = "binomial")
plot(result)
summary(result)
}
