% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_wrappers.R
\name{fastcpd_family}
\alias{fastcpd_family}
\alias{fastcpd.family}
\title{Wrapper functions for fastcpd}
\description{
Wrapper functions for fastcpd to find change points in various
models.
}
\seealso{
\code{\link[=fastcpd.mean]{fastcpd.mean()}}, \code{\link[=fastcpd.variance]{fastcpd.variance()}}, \code{\link[=fastcpd.mv]{fastcpd.mv()}},
\code{\link[=fastcpd.meanvariance]{fastcpd.meanvariance()}} for basic statistics change models;
\code{\link[=fastcpd.lm]{fastcpd.lm()}}, \code{\link[=fastcpd.binomial]{fastcpd.binomial()}}, \code{\link[=fastcpd.poisson]{fastcpd.poisson()}},
\code{\link[=fastcpd.lasso]{fastcpd.lasso()}} for regression coefficients change models;
\code{\link[=fastcpd.ar]{fastcpd.ar()}}, \code{\link[=fastcpd.var]{fastcpd.var()}}, \code{\link[=fastcpd.arima]{fastcpd.arima()}}, \code{\link[=fastcpd.arma]{fastcpd.arma()}},
\code{\link[=fastcpd.garch]{fastcpd.garch()}} for change in time series models.
}
\keyword{internal}
