% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_wrappers.R
\name{fastcpd.meanvariance}
\alias{fastcpd.meanvariance}
\alias{fastcpd_meanvariance}
\alias{fastcpd.mv}
\alias{fastcpd_mv}
\title{Find change points efficiently in variance change models}
\usage{
fastcpd.meanvariance(data, ...)

fastcpd_meanvariance(data, ...)

fastcpd.mv(data, ...)

fastcpd_mv(data, ...)
}
\arguments{
\item{data}{A matrix, a data frame or a vector.}

\item{...}{Other arguments passed to \code{\link{fastcpd}}, for example,
\code{segment_count}.}
}
\value{
A class \code{fastcpd} object.
}
\description{
\code{fastcpd_meanvariance}, \code{fastcpd.meanvariance},
\code{fastcpd_mv}, \code{fastcpd.mv} are wrapper
functions of \code{\link{fastcpd}} to find the meanvariance change. The
function is similar to \code{\link{fastcpd}} except that the data is by
default a matrix or data frame or a vector with each row / element as an
observation and thus a formula is not required here.
}
\examples{
set.seed(1)
p <- 3
result <- fastcpd.mv(
  rbind(
    mvtnorm::rmvnorm(300, mean = rep(0, p), sigma = diag(1, p)),
    mvtnorm::rmvnorm(400, mean = rep(10, p), sigma = diag(1, p)),
    mvtnorm::rmvnorm(300, mean = rep(0, p), sigma = diag(50, p)),
    mvtnorm::rmvnorm(300, mean = rep(0, p), sigma = diag(1, p)),
    mvtnorm::rmvnorm(400, mean = rep(10, p), sigma = diag(1, p)),
    mvtnorm::rmvnorm(300, mean = rep(10, p), sigma = diag(50, p))
  )
)
summary(result)
}
