% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_ts.R
\name{fastcpd.ts}
\alias{fastcpd.ts}
\alias{fastcpd_ts}
\title{Find change points efficiently in time series data}
\usage{
fastcpd.ts(data, family = NULL, order = NULL, ...)

fastcpd_ts(data, family = NULL, order = NULL, ...)
}
\arguments{
\item{data}{A numeric vector, a matrix, a data frame or a time series object.}

\item{family}{A character string specifying the family of the time series.
The value should be one of \code{"ar"} or \code{"var"}.}

\item{order}{A positive integer specifying the order of the time series.}

\item{...}{Other arguments passed to \code{\link{fastcpd}}.}
}
\value{
A class \code{fastcpd} object.
}
\description{
`fastcpd_ts` is a wrapper function for `fastcpd` to find
  change points in time series data. The function is similar to `fastcpd`
  except that the data is a time series data and the family is one of
  \code{"ar"} or \code{"var"}.
}
