% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_log_lik.R
\name{neg_log_lik}
\alias{neg_log_lik}
\title{Solve logistic regression using Gradient Descent Extension to the
multivariate case}
\usage{
neg_log_lik(data, b)
}
\arguments{
\item{data}{TODO}

\item{b}{TODO}
}
\value{
TODO
}
\description{
Solve logistic regression using Gradient Descent Extension to the
multivariate case
}
