% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_log_likelihood.R
\name{negative_log_likelihood}
\alias{negative_log_likelihood}
\title{Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case}
\usage{
negative_log_likelihood(data, theta, family, lambda)
}
\arguments{
\item{data}{A data frame containing the data to be segmented.}

\item{theta}{Estimate of the parameters. If null, the function will estimate
the parameters.}

\item{family}{Family of the model.}

\item{lambda}{Lambda for L1 regularization. Only used for lasso.}
}
\value{
Negative log likelihood of the corresponding data with the given
  family.
}
\description{
Solve logistic/poisson regression using Gradient Descent Extension to the
multivariate case
}
