% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_class.R
\name{plot.fastcpd}
\alias{plot.fastcpd}
\alias{plot,fastcpd,missing-method}
\title{Plot the data and the change points for a \code{fastcpd} object}
\usage{
\method{plot}{fastcpd}(x, ...)

\S4method{plot}{fastcpd,missing}(x, y, ...)
}
\arguments{
\item{x}{\code{fastcpd} object.}

\item{...}{Ignored.}

\item{y}{Ignored.}
}
\value{
No return value, called for plotting.
}
\description{
Plot the data and the change points for a \code{fastcpd} object
}
