% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_class.R
\name{print.fastcpd}
\alias{print.fastcpd}
\alias{print,fastcpd-method}
\title{Print the call and the change points for a \code{fastcpd} object}
\usage{
\method{print}{fastcpd}(x, ...)

\S4method{print}{fastcpd}(x, ...)
}
\arguments{
\item{x}{\code{fastcpd} object.}

\item{...}{Ignored.}
}
\value{
Return a (temporarily) invisible copy of the \code{fastcpd} object.
  Called primarily for printing the change points in the model.
}
\description{
Print the call and the change points for a \code{fastcpd} object
}
