% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.fastcpd}
\alias{print.fastcpd}
\alias{print,fastcpd-method}
\title{Print the call and the change points for a \linkS4class{fastcpd} object}
\usage{
\method{print}{fastcpd}(x, ...)

\S4method{print}{fastcpd}(x, ...)
}
\arguments{
\item{x}{A \linkS4class{fastcpd} object.}

\item{...}{Ignored.}
}
\value{
Return a (temporarily) invisible copy of the \linkS4class{fastcpd} object.
Called primarily for printing the change points in the model.
}
\description{
Print the call and the change points for a \linkS4class{fastcpd} object
}
