% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcpd_data.R
\docType{data}
\name{well_log}
\alias{well_log}
\title{Well-log Dataset from Numerical Bayesian Methods Applied to Signal Processing}
\format{
A Time-Series of length 3989.
}
\source{
<https://github.com/alan-turing-institute/TCPD>
}
\usage{
well_log
}
\description{
This is the well-known well-log dataset used in many changepoint papers
obtained from Alan Turing Institute GitHub repository and licensed under
the MIT license. Outliers with value less or equal to 1e5 are removed.
}
\examples{
result <- fastcpd.mean(well_log, trim = 0.002)
summary(result)
plot(result)
}
\keyword{datasets}
